requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.6.6
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.7
django-import-export==2.6.1
django-widget-tweaks==1.4.9
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
isort==5.9.3
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.8.2
PyMySQL==0.10.0
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-decouple==3.5
pytz==2020.4
PyYAML==6.0
reportlab==3.5.55
six==1.15.0
sqlparse==0.4.1
tablib==3.1.0
tinycss2==1.1.0
toml==0.10.2
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
